Change DEBUG to True in local-settings.py
This is because django will not serve static files in production (DEBUG=False). This should be the job of the web server (in this case Nginx). Since the local registry isn't using a proper web server and is HTTP only, we are already in "grossly inefficient and probably insecure" territory so having DEBUG set to 'True' seems like the best option right now.